requirements.txt
httpx
aiogram
python-dotenv
requests